refactor(mysql): split CLI process lifecycle by execution path
